+++ src/test/java/com/tiketeer/Tiketeer/testhelper/TestHelper.java
@@ -13,7 +13,7 @@ import com.tiketeer.Tiketeer.domain.member.Member;
 import com.tiketeer.Tiketeer.domain.member.Otp;
 import com.tiketeer.Tiketeer.domain.member.repository.MemberRepository;
 import com.tiketeer.Tiketeer.domain.member.repository.OtpRepository;
-import com.tiketeer.Tiketeer.domain.member.service.LoginService;
+import com.tiketeer.Tiketeer.domain.member.usecase.LoginUseCase;
 import com.tiketeer.Tiketeer.domain.member.usecase.dto.LoginCommandDto;
 import com.tiketeer.Tiketeer.domain.member.usecase.dto.LoginResultDto;
 import com.tiketeer.Tiketeer.domain.purchase.repository.PurchaseRepository;
@@ -39,7 +39,7 @@ public class TestHelper {
 	private final TicketRepository ticketRepository;
 	private final TicketingRepository ticketingRepository;
 	private final PasswordEncoder passwordEncoder;
-	private final LoginService loginService;
+	private final LoginUseCase loginUseCase;
 
 	@Autowired
 	public TestHelper(
@@ -52,7 +52,7 @@ public class TestHelper {
 		TicketRepository ticketRepository,
 		TicketingRepository ticketingRepository,
 		PasswordEncoder passwordEncoder,
-		LoginService loginService
+		LoginUseCase loginUseCase
 	) {
 		this.permissionRepository = permissionRepository;
 		this.roleRepository = roleRepository;
@@ -63,7 +63,7 @@ public class TestHelper {
 		this.ticketRepository = ticketRepository;
 		this.ticketingRepository = ticketingRepository;
 		this.passwordEncoder = passwordEncoder;
-		this.loginService = loginService;
+		this.loginUseCase = loginUseCase;
 	}
 
 	@Transactional
@@ -102,7 +102,7 @@ public class TestHelper {
 	public String registerAndLoginAndReturnAccessToken(String email, RoleEnum roleEnum) {
 		var password = "1q2w3e4r!!";
 		createMember(email, "1q2w3e4r!!", roleEnum);
-		return loginService.login(LoginCommandDto.builder().email(email).password(password).build()).getAccessToken();
+		return loginUseCase.login(LoginCommandDto.builder().email(email).password(password).build()).getAccessToken();
 	}
 
 	@Transactional
@@ -115,7 +115,7 @@ public class TestHelper {
 	public LoginResultDto registerAndLoginAndReturnAccessTokenAndRefreshToken(String email, RoleEnum roleEnum) {
 		String password = "1q2w3e4r!!";
 		createMember(email, password, roleEnum);
-		return loginService.login(LoginCommandDto.builder().email(email).password(password).build());
+		return loginUseCase.login(LoginCommandDto.builder().email(email).password(password).build());
 	}
 
 	@Transactional
